test: split profiler tests into fixtures and test modules
