+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+addopts = -q
+testpaths = tests
