+++ tests/conftest.py
@@ -1,22 +1,25 @@
+from collections.abc import Iterator
 from typing import Any
 
 import pytest
-from sqlalchemy import create_engine, text
+from sqlalchemy import create_engine
+from sqlalchemy.engine import Engine
 
 from fastapi_silk.profiler import setup_sql_profiler
 from fastapi_silk.storage import request_queries
+from tests.fixtures.sql_queries import SQL_QUERIES
 
 
 @pytest.fixture
-def engine():
+def engine() -> Engine:
     """Create an in-memory SQLite engine with SQL profiler attached."""
-    engine = create_engine("sqlite:///:memory:")
-    setup_sql_profiler(engine)
-    return engine
+    db_engine = create_engine("sqlite:///:memory:")
+    setup_sql_profiler(db_engine)
+    return db_engine
 
 
 @pytest.fixture(autouse=True)
-def reset_queries() -> None:
+def reset_queries() -> Iterator[None]:
     """Ensure query storage is clean before and after each test."""
     request_queries.set([])
     try:
@@ -25,27 +28,13 @@ def reset_queries() -> None:
         request_queries.set([])
 
 
-@pytest.fixture(params=["SELECT 1"])
+@pytest.fixture(params=SQL_QUERIES)
 def sql_query(request: pytest.FixtureRequest) -> str:
     """SQL statements to execute, easily extendable/parameterized."""
-    return request.param
+    return str(request.param)
 
 
 @pytest.fixture
 def queries() -> list[dict[str, Any]]:
     """Expose captured SQL queries for assertions."""
     return request_queries.get()
-
-
-def test_profiler_captures_query(
-    engine,
-    sql_query: str,
-    queries: list[dict[str, Any]],
-) -> None:
-    """Profiler should intercept queries and store SQL and duration."""
-    with engine.connect() as conn:
-        conn.execute(text(sql_query))
-
-    assert len(queries) == 1
-    assert sql_query in str(queries[0]["sql"])
-    assert "duration" in queries[0]
