+++ tests/fixtures/sql_queries.py
@@ -1,0 +1,3 @@
+SQL_QUERIES = [
+    "SELECT 1",
+]
